FInished a lot of work
Added a print button to the final page
Changed all the front end UI
Updated the movie database with more movies

There's an issue with movies. We only had movies for one theater so I added more movies for each theater. That is fine, but for the show Times page it displays the times for that given movie for all theaters. It needs to ONLY show times for that movie AND theater.